+++ src/devices/colorbulb.ts
@@ -1,8 +1,8 @@
-import { Service, PlatformAccessory, CharacteristicValue, MacAddress } from 'homebridge';
+import { Service, PlatformAccessory, CharacteristicValue } from 'homebridge';
 import { SwitchBotPlatform } from '../platform';
 import { interval, Subject } from 'rxjs';
 import { debounceTime, skipWhile, tap } from 'rxjs/operators';
-import { DeviceURL, device, devicesConfig } from '../settings';
+import { DeviceURL, device, devicesConfig, switchbot, deviceStatusResponse } from '../settings';
 import { AxiosResponse } from 'axios';
 
 /**
@@ -19,33 +19,27 @@ export class ColorBulb {
   Brightness!: CharacteristicValue;
   ColorTemperature!: CharacteristicValue;
 
-  // Others
-  deviceStatus: any;
-  set_minStep: any;
-  switchbot!: {
-    discover: (
-      arg0:
-        {
-          duration: any;
-          model: string;
-          quick: boolean;
-          id: MacAddress;
-        }
-    ) => Promise<any>;
-    wait: (
-      arg0: number
-    ) => any;
-  };
+  // OpenAPI Others
+  deviceStatus!: deviceStatusResponse;
+
+  // BLE Others
+  switchbot!: switchbot;
+
+  // Config
+  set_minStep?: number;
 
   // Updates
   colorBulbUpdateInProgress!: boolean;
-  doColorBulbUpdate;
+  doColorBulbUpdate!: Subject<void>;
 
   constructor(
     private readonly platform: SwitchBotPlatform,
     private accessory: PlatformAccessory,
     public device: device & devicesConfig,
   ) {
+    // ColorBulb Config
+    this.platform.device(`[ColorBulb Config] ble: ${device.ble}, set_minStep: ${device.colorbulb?.set_minStep}`);
+
     // default placeholders
     this.On = false;
     this.Brightness = 0;