+++ src/devices/curtains.ts
@@ -20,31 +20,18 @@ export class Curtain {
   BatteryLevel?: CharacteristicValue;
   StatusLowBattery?: CharacteristicValue;
 
+  // OpenAPI Others
+  deviceStatus!: deviceStatusResponse;
+  setNewTarget!: boolean;
+  setNewTargetTimer!: NodeJS.Timeout;
+
   // BLE Others
   switchbot!: switchbot;
   serviceData!: serviceData;
-  moveTimer!: NodeJS.Timeout;
-  ScanIntervalId!: NodeJS.Timeout;
-  setNewTargetTimer!: NodeJS.Timeout;
-  AutoDisableFastScanTimeoutId!: NodeJS.Timeout;
-  setNewTarget!: boolean;
-  moveTime!: number;
-  ScanDuration: any;
-  ReverseDir: any;
-  FastScanEnabled!: boolean;
-  FastScanInterval!: number;
-  SlowScanInterval!: number;
-  OpenCloseThreshold: any;
-  PreviousPosition: any;
-  Position: any;
-  FastScanDuration!: number;
   calibration: serviceData['calibration'];
   battery: serviceData['battery'];
   position: serviceData['position'];
   lightLevel: serviceData['lightLevel'];
-
-  // OpenAPI Others
-  deviceStatus!: deviceStatusResponse;
 
   // Config
   set_minStep!: number;
@@ -59,12 +46,15 @@ export class Curtain {
     private accessory: PlatformAccessory,
     public device: device & devicesConfig,
   ) {
+    // Curtain Config
+    this.platform.device(`[Curtain Config] ble: ${device.ble}, disable_group: ${device.curtain?.disable_group},`
+    + ` refreshRate: ${device.curtain?.refreshRate}, set_max: ${device.curtain?.set_max}, set_min: ${device.curtain?.set_min},`
+    + ` set_minStep: ${device.curtain?.set_minStep}`);
+
     // default placeholders
-    this.setMinMax();
     this.CurrentPosition = 0;
     this.TargetPosition = 0;
     this.PositionState = this.platform.Characteristic.PositionState.STOPPED;
-    this.ScanDuration = this.platform.config.options!.refreshRate!;
 
     // BLE Connection
     if (device.ble) {
@@ -98,8 +88,6 @@ export class Curtain {
       accessory.getService(this.platform.Service.WindowCovering) ||
       accessory.addService(this.platform.Service.WindowCovering)), `${device.deviceName} ${device.deviceType}`;
 
-    this.platform.device(`[Curtain Config] disable_group: ${device.curtain?.disable_group}, refreshRate: ${device.curtain?.refreshRate},`
-      + ` set_max: ${device.curtain?.set_max}, set_min: ${device.curtain?.set_min}, set_minStep: ${device.curtain?.set_minStep}`);
     // To avoid "Cannot add a Service with the same UUID another Service without also defining a unique 'subtype' property." error,
     // when creating multiple services of the same type, you need to use the following syntax to specify a name and subtype id:
     // accessory.getService('NAME') ?? accessory.addService(this.platform.Service.Lightbulb, 'NAME', 'USER_DEFINED_SUBTYPE');
@@ -123,7 +111,7 @@ export class Curtain {
         validValueRanges: [0, 100],
       })
       .onGet(() => {
-        return Number.isNaN(this.CurrentPosition);
+        return this.CurrentPosition;
       });
 
     this.service
@@ -217,18 +205,31 @@ export class Curtain {
 
   private async BLEparseStatus() {
     this.platform.device('Curtains BLE Device parseStatus');
+    this.CurrentPosition = 100 - Number(this.position);
     this.setMinMax();
-    this.CurrentPosition = Number(this.position);
-    this.setMinMax();
-    if (this.TargetPosition > this.CurrentPosition) {
-      this.platform.debug(`Curtain ${this.accessory.displayName} - Current position: ${this.CurrentPosition} closing`);
-      this.PositionState = this.platform.Characteristic.PositionState.INCREASING;
-    } else if (this.TargetPosition < this.CurrentPosition) {
-      this.platform.debug(`Curtain ${this.accessory.displayName} - Current position: ${this.CurrentPosition} opening`);
-      this.PositionState = this.platform.Characteristic.PositionState.DECREASING;
+    this.platform.debug(`Curtain ${this.accessory.displayName} CurrentPosition - Device is Currently: ${this.CurrentPosition}`);
+    if (this.setNewTarget) {
+      this.platform.log.info(`Checking ${this.accessory.displayName} Status ...`);
+    }
+
+    if (this.setNewTarget) {
+      if (this.TargetPosition > this.CurrentPosition) {
+        this.platform.debug(`Curtain ${this.accessory.displayName} - Current position: ${this.CurrentPosition} closing`);
+        this.PositionState = this.platform.Characteristic.PositionState.INCREASING;
+      } else if (this.TargetPosition < this.CurrentPosition) {
+        this.platform.debug(`Curtain ${this.accessory.displayName} - Current position: ${this.CurrentPosition} opening`);
+        this.PositionState = this.platform.Characteristic.PositionState.DECREASING;
+      } else {
+        this.platform.debug(`Curtain ${this.CurrentPosition} - standby`);
+        this.PositionState = this.platform.Characteristic.PositionState.STOPPED;
+      }
     } else {
-      this.platform.debug(`Curtain ${this.CurrentPosition} - standby`);
-      this.PositionState = this.platform.Characteristic.PositionState.STOPPED;
+      this.platform.debug(`Curtain ${this.accessory.displayName} - Current position: ${this.CurrentPosition} standby`);
+      if (!this.setNewTarget) {
+        /*If Curtain calibration distance is short, there will be an error between the current percentage and the target percentage.*/
+        this.TargetPosition = this.CurrentPosition;
+        this.PositionState = this.platform.Characteristic.PositionState.STOPPED;
+      }
     }
     // Brightness
     switch (this.lightLevel) {
@@ -277,7 +278,7 @@ export class Curtain {
   }
 
   private async openAPIparseStatus() {
-    this.setMinMax();
+    this.platform.device('Curtains OpenAPI Device parseStatus');
     this.CurrentPosition = 100 - this.deviceStatus.body.slidePosition!;
     this.setMinMax();
     this.platform.debug(`Curtain ${this.accessory.displayName} CurrentPosition - Device is Currently: ${this.CurrentPosition}`);
@@ -376,10 +377,10 @@ export class Curtain {
   }
 
   private async openAPIRefreshStatus() {
+    this.platform.debug('Curtains OpenAPI Device RefreshStatus');
     try {
       this.deviceStatus = (await this.platform.axios.get(`${DeviceURL}/${this.device.deviceId}/status`)).data;
       this.platform.debug(`Curtain ${this.accessory.displayName} refreshStatus: ${JSON.stringify(this.deviceStatus)}`);
-      this.setMinMax();
       this.parseStatus();
       this.updateHomeKitCharacteristics();
     } catch (e: any) {
@@ -404,7 +405,7 @@ export class Curtain {
     const switchbot = this.connectBLE();
     switchbot.discover({ model: 'c', quick: true, id: this.device.bleMac }).then((device_list) => {
       this.platform.log.info(`${this.accessory.displayName}, Target Position: ${this.TargetPosition}`);
-      return device_list[0].runToPos(Number(this.TargetPosition));
+      return device_list[0].runToPos(100 - Number(this.TargetPosition));
     }).then(() => {
       this.platform.device('Done.');
     }).catch((e: any) => {
@@ -450,9 +451,6 @@ export class Curtain {
         CurrentAmbientLightLevel: this.CurrentAmbientLightLevel,
       })}`,
     );
-    if (!this.device.ble) {
-      this.setMinMax();
-    }
     if (this.CurrentPosition === undefined) {
       this.platform.debug(`Curtain ${this.accessory.displayName} CurrentPosition: ${this.CurrentPosition}`);
     } else {
@@ -504,7 +502,7 @@ export class Curtain {
     }
   }
 
-  private statusCode(push: AxiosResponse<{ statusCode: number;}>) {
+  private statusCode(push: AxiosResponse<{ statusCode: number; }>) {
     switch (push.data.statusCode) {
       case 151:
         this.platform.log.error('Command not supported by this device type.');
@@ -540,18 +538,16 @@ export class Curtain {
 
     this.TargetPosition = value;
 
+    this.setMinMax();
     if (value > this.CurrentPosition) {
       this.PositionState = this.platform.Characteristic.PositionState.INCREASING;
       this.setNewTarget = true;
-      this.setMinMax();
     } else if (value < this.CurrentPosition) {
       this.PositionState = this.platform.Characteristic.PositionState.DECREASING;
       this.setNewTarget = true;
-      this.setMinMax();
     } else {
       this.PositionState = this.platform.Characteristic.PositionState.STOPPED;
       this.setNewTarget = false;
-      this.setMinMax();
     }
     this.service.setCharacteristic(this.platform.Characteristic.PositionState, this.PositionState);